% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_bathymetry}
\alias{get_bathymetry}
\title{NOAA ETOPO 15-arcsecond digitial bedrock elevation model for the EBS}
\source{
NOAA National Centers for Environmental Information. 2022: ETOPO 2022 15 Arc-Second Global Relief Model. NOAA National Centers for Environmental Information. \href{https://doi.org/10.25921/fd45-gt74}{doi.org/10.25921/fd45-gt74}
}
\usage{
get_bathymetry()
}
\value{
A "stars" object
}
\description{
NOAA ETOPO 15-arcsecond digitial bedrock elevation model for the EBS
}
