% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_ebs_shapefile}
\alias{get_ebs_shapefile}
\title{EBS survey shapefiles}
\source{
\href{https://github.com/afsc-gap-products/akgfmaps}{github.com/afsc-gap-products/akgfmaps}
}
\usage{
get_ebs_shapefile(region = c("EBS", "SEBS"), type = c("boundary", "grid"))
}
\arguments{
\item{region}{Either "EBS" for full survey region including NBS, or "SEBS"}

\item{type}{Whether to return multipolygon corresponding to survey grid ("grid") or region boundary only ("boundary")}
}
\value{
An "sf" object
}
\description{
EBS survey shapefiles
}
