% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mom6.R
\name{get_mom6}
\alias{get_mom6}
\title{Download gridded estimates from CEFI MOM6 experiments}
\usage{
get_mom6(
  var = "tob",
  freq = c("monthly", "daily"),
  grid_type = c("regrid", "raw"),
  region = c("NEP", "NWA"),
  experiment = c("hindcast", "forecast", "reforecast", "decadal_forecast"),
  category = paste0("ocean_", freq),
  release = NA,
  extent = NULL,
  start_date = NA,
  end_date = NA,
  target_crs = NULL,
  chunk = "none"
)
}
\arguments{
\item{var}{variable name. See [CEFI portal](https://psl.noaa.gov/cefi_portal/) or use
`list_mom6()` for options.}

\item{freq}{"monthly" or "daily". Note that a matching category must also be
specified, e.g. "ocean_daily" for daily frequency.}

\item{grid_type}{"regrid" or "raw". Regridded data is on a regular grid, while raw data is on the model's native curvilinear grid.}

\item{region}{CEFI region. "NEP" for Northeast Pacific or "NWA" for Northwest Atlantic.}

\item{experiment}{CEFI experiment. Defaults to "hindcast".}

\item{category}{data category. See [CEFI portal](https://psl.noaa.gov/cefi_portal/) or use
`list_mom6()` for options.}

\item{release}{release code. If NA, returns latest available release.}

\item{extent}{Either (1) a shapefile from which to compute the extent,
(2) a bounding box created using `sf::st_bbox()` with accompanying CRS,
or (3) NA or NULL for the full MOM6 grid. If a shapefile is provided, it will be
used to mask (crop) the output prior to returning. For the NEP region, defaults
to the NOAA AFSC Eastern Bering Sea (including NBS) survey region. For NWA, defaults to NA.}

\item{start_date}{Initial date to query data for. If frequency = "monthly",
only years and months of provided date is used, date of month is ignored.
If NA, all available dates are returned. Can be NA even if `end_date` is provided.}

\item{end_date}{Final date to query data for. If frequency = "monthly",
only years and months of provided date is used, date of month is ignored.
If NA, all available dates are returned. Can be NA even if `start_date` is provided.}

\item{target_crs}{CRS to transform output to (as `crs` object or integer EPSG code). 
Defaults to UTM zone 2N (EPSG 32602) for NEP regridded data. NA or NULL defaults to 
WGS84 lat/long (EPSG 4326).}

\item{chunk}{Time interval to chunk requests by (e.g., "10 years"); "none" for no chunking.
Using no chunks may result in request hitting server data limits; using too many chunks 
may result in hitting server rate limits. Accepts any string that `seq.Date` can parse.}
}
\value{
A `stars` object
}
\description{
Download gridded estimates from CEFI MOM6 experiments
}
