% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mom6_nep.R
\name{get_mom6_nep}
\alias{get_mom6_nep}
\title{Download gridded estimates from the CEFI MOM6 NEP hindcast}
\usage{
get_mom6_nep(
  var = "tob",
  freq = c("monthly", "daily"),
  category = paste0("ocean_", freq),
  release = NA,
  extent = get_ebs_shapefile("EBS"),
  start_date = NA,
  end_date = NA,
  target_crs = 32602,
  chunk = "none"
)
}
\arguments{
\item{var}{variable name. See [CEFI portal](https://psl.noaa.gov/cefi_portal/) or use
`list_mom6_nep()` for options.}

\item{freq}{"monthly" or "daily". Note that a matching category must also be
specified, e.g. "ocean_daily" for daily frequency.}

\item{category}{data category. See [CEFI portal](https://psl.noaa.gov/cefi_portal/) or use
`list_mom6_nep()` for options.}

\item{release}{release code. If NA, returns latest available release.}

\item{extent}{Either (1) a shapefile from which to compute the extent,
(2) a bounding box created using `sf::st_bbox()` with accompanying CRS,
or (3) NA for the full MOM6 NEP grid. If a shapefile is provided, it will be
used to mask (crop) the output prior to returning. Defaults to NOAA AFSC
Eastern Bering Sea (including NBS) survey region.}

\item{start_date}{Initial date to query data for. If frequency = "monthly",
only years and months of provided date is used, date of month is ignored.
If NA, all available dates are returned. Can be NA even if `end_date` is provided.}

\item{end_date}{Final date to query data for. If frequency = "monthly",
only years and months of provided date is used, date of month is ignored.
If NA, all available dates are returned. Can be NA even if `start_date` is provided.}

\item{target_crs}{CRS to transform output to (as `crs` object or integer EPSG code). 
Defaults to UTM zone 2N (EPSG 32602). NA defaults to WGS84 lat/long (EPSG 4326).}

\item{chunk}{Time interval to chunk requests by (e.g., "10 years"); "none" for no chunking.
Using no chunks may result in request hitting server data limits; using too many chunks 
may result in hitting server rate limits. Accepts any string that `seq.Date` can parse.}
}
\value{
A `stars` object
}
\description{
Download gridded estimates from the CEFI MOM6 NEP hindcast
}
