% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mom6.R
\name{list_mom6}
\alias{list_mom6}
\title{List available CEFI MOM6 datasets}
\usage{
list_mom6(
  region = c("NEP", "NWA"),
  experiment = c("hindcast", "forecast", "reforecast", "decadal_forecast"),
  var = NULL,
  freq = NULL,
  category = NULL,
  release = NULL
)
}
\arguments{
\item{region}{MOM6 model region ("NEP" or "NWA")}

\item{experiment}{Model experiment ("hindcast", "forecast", "reforecast", "decadal_forecast")}

\item{var}{(optional) variable name; will return all variables if missing}

\item{freq}{(optional) frequency ("daily" or "monthly")}

\item{category}{(optional) CEFI data category}

\item{release}{(optional) CEFI release code}
}
\value{
A `data.frame` with available datasets
}
\description{
List available CEFI MOM6 datasets
}
