% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mom6_nep.R
\name{list_mom6_nep}
\alias{list_mom6_nep}
\title{List available CEFI MOM6 NEP datasets}
\usage{
list_mom6_nep(var = NULL, freq = NULL, category = NULL, release = NULL)
}
\arguments{
\item{var}{(optional) variable name; will return all variables if missing}

\item{freq}{(optional) frequency ("daily" or "monthly")}

\item{category}{(optional) CEFI data category}

\item{release}{(optional) CEFI release code}
}
\value{
A `data.frame` with available datasets
}
\description{
List available CEFI MOM6 NEP datasets
}
