% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_sediment}
\alias{get_sediment}
\title{Log sediment grain size (Krumbein phi scale) for the EBS region}
\source{
Richwine, K. A. et al. 2018. Surficial sediments of the eastern Bering Sea continental shelf: EBSSED-2 database documentation. – US Dept Comm, NOAA Tech Mem. NMFS-AFSC-377
}
\usage{
get_sediment()
}
\value{
A "stars" object
}
\description{
Estimates from the EBSSED database interpolated / extrapolated using
ordinary kriging with an exponential fit to the empirical semi-variogram.
}
